+++ src/Exceptions/CallableNotFoundException.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * phpgram
+ *
+ * This File is part of the phpgram Micro Framework
+ *
+ * Web: https://gitlab.com/grammm/php-gram/phpgram
+ *
+ * @license https://gitlab.com/grammm/php-gram/phpgram/blob/master/LICENSE
+ *
+ * @author Jörn Heinemann <joernheinemann@gmx.de>
+ */
+
+namespace Gram\Exceptions;
+
+
+class CallableNotFoundException extends AbstractGramException
+{
+
+}